+++ packages/contracts/test/utils/deployer.ts
@@ -4,9 +4,12 @@ import * as path from 'path';
 
 import { constants } from '../../util/constants';
 
+import { web3 } from './web3_wrapper';
+
 const deployerOpts = {
+    web3Provider: web3.currentProvider,
     artifactsDir: path.resolve('src', 'artifacts'),
-    jsonrpcPort: devConstants.RPC_PORT,
+    jsonrpcUrl: devConstants.RPC_URL,
     networkId: constants.TESTRPC_NETWORK_ID,
     defaults: {
         gas: devConstants.GAS_ESTIMATE,