+++ packages/deployer/src/globals.d.ts
@@ -1,11 +1,51 @@
+// tslint:disable:completed-docs
 declare module 'solc' {
-    // tslint:disable:completed-docs
-    export function compile(sources: any, optimizerEnabled: number, findImports: (importPath: string) => any): any;
-    export function setupMethods(solcBin: any): any;
-    // tslint:enable:completed-docs
+    import * as Web3 from 'web3';
+    export interface ContractCompilationResult {
+        srcmap: string;
+        srcmapRuntime: string;
+        bytecode: string;
+        runtimeBytecode: string;
+        interface: string;
+    }
+    export interface CompilationResult {
+        errors: string[];
+        contracts: {
+            [contractIdentifier: string]: ContractCompilationResult;
+        };
+        sources: {
+            [sourceName: string]: {
+                AST: any;
+            };
+        };
+        sourceList: string[];
+    }
+    export interface ImportContents {
+        contents: string;
+    }
+    export interface InputSources {
+        sources: {
+            [fileName: string]: string;
+        };
+    }
+    export interface SolcInstance {
+        compile(
+            sources: InputSources,
+            optimizerEnabled: number,
+            findImports: (importPath: string) => ImportContents,
+        ): CompilationResult;
+    }
+    export function loadRemoteVersion(versionName: string, cb: (err: Error | null, res?: SolcInstance) => void): void;
+    export function setupMethods(solcBin: any): SolcInstance;
 }
 
 declare module 'web3-eth-abi' {
-    // tslint:disable-next-line:completed-docs
     export function encodeParameters(typesArray: string[], parameters: any[]): string;
+}
+
+declare module '*.json' {
+    const json: any;
+    /* tslint:disable */
+    export default json;
+    /* tslint:enable */
 }