+++ packages/react-docs/src/utils/utils.ts
@@ -1,0 +1,5 @@
+export const utils = {
+    spawnSwitchErr(name: string, value: any) {
+        return new Error(`Unexpected switch value: ${value} encountered for ${name}`);
+    },
+};