+++ packages/sra-report/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## v0.0.2 - _TBD, 2018_
 
     * Added postman collection unit tests (#451)
+    * Consolidate all `console.log` calls into `logUtils` in the `@0xproject/utils` package (#452)
 
 ## v0.0.1 - _March 8, 2018_