+++ packages/subproviders/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.8.0 - _TBD, 2018_
+
+    * Export `GanacheSubprovider` and `Subprovider` (#426)
+    * Make all subproviders to derive from `Subprovider` (#426)
+    * Add types for `NextCallback`, `OnNextCompleted` (#426)
+
 ## v0.7.0 - _March 8, 2018_
 
     * Updated legerco packages. Removed node-hid package as a dependency and make it an optional dependency. It is still used in integration tests but is causing problems for users on Linux distros. (#437)